==> NonEnglish Categorical Trait Prediction/01. Description_Preprocessing.ipynb <==
**Notebook to preprocess the Spanish and German Wikipedia descriptions.**

This is done by removing unrelated information and processing the text using NLP using the following pipeline:
* Remove unnecessary information such as description categories that are not relevant to the task
* Remove artifacts from text
* Preprocess by lowercasing and tokenizing, optionally removing numbers, punctuation and stopwords

# Libraries & Functions

'''Math & Data Libraries'''
import numpy as np
import pandas as pd

''' Miscellaneous Libraries'''
from tqdm import tqdm

'''NLP Libraries'''
import nltk
from nltk.corpus import stopwords
import string

import unidecode
import unicodedata

nltk.download('punkt')
nltk.download('stopwords')

## NLP Functions

def remove_artifacts(text):
    """
    Remove artifacts from text. 
    ---
    Parameters
    ----------
    text : str
        text from which artifacts should be removed

    Returns
    -------
    text_removed_artifacts : str
        text with removed artifacts
    """
    return str(text).replace("nbsp", "").replace("<p>", "").replace("</p>", "").replace("<i>", "").replace("</i>", "").replace("<b>", "").replace("</b>", "").replace("&","").replace("_x000D_", "").replace("\"", "").replace("\'", "").replace("<em>", "").replace("</em>", "").replace("<br>", "").replace("</br>", "")

def preprocess_description(text, remove_numbers = True, remove_stopwords = False):
    """
    Preprocess text with the following steps:
        1. Lowercasing
        2. Removal of digits (if remove_numbers = True)
        3. Removal of punctuation (if remove_numbers = True)
        4. Removal of stopwords (if remove_stopwords = True) 
        5. Tokenization and join to fix issues with blank space
    ---
    Parameters
    ----------
    text : str
        text from which artifacts should be removed
    remove_numbers : boolean
        whether to remove or keep digits in the text
    remove_stopwords : boolean
        whether to remove or stopwords in the text

    Returns
    -------
    preprocessed_text : str
        preprocessed_text
    """
    text = text.lower()
    if remove_numbers:
        text = text.translate(str.maketrans('', '', string.digits))
        text = text.translate(str.maketrans('', '', string.punctuation))
       
    tokens = nltk.word_tokenize(text)
    text = " ".join(tokens)

    if remove_numbers:
        tokens_long = []
        for token in tokens:
            if len(token)>1:
                tokens_long.append(token)
        text = ' '.join(tokens_long)

    if remove_stopwords:
        stopwords_ = stopwords.words('english')
        tokens = nltk.word_tokenize(text)
        tokens_stop_words = []
        for token in tokens:
            if token not in stopwords_:
                tokens_stop_words.append(token)
        text = ' '.join(tokens_stop_words)

    return text

# Input Data

## Spanish Wikipedia

df_WIKI_ES = pd.read_excel("..//Datasets//Initial Databases//WIKI_orig_descriptions_ESP.xlsx")

df_WIKI_ES

print("Initial Number of Categories:", df_WIKI_ES["WIKI_id"].nunique())

## German Wikipedia

df_WIKI_DE = pd.read_excel("..//Datasets//Initial Databases//WIKI_orig_descriptions_DE.xlsx")

df_WIKI_DE

print("Initial Number of Categories:", df_WIKI_DE["WIKI_id"].nunique())

# Preprocessing

## 1. Remove artifacts and accents from text
Such as $<p>, \&nbsp, <i>, <b>, <br>, <em>, _x000D_, ", ', $

df_WIKI_ES.loc[:, "prep_description_1"] = df_WIKI_ES["description"].apply(remove_artifacts)
df_WIKI_ES.loc[:, "prep_description_1"] = df_WIKI_ES["prep_description_1"].apply(unidecode.unidecode)

df_WIKI_DE.loc[:, "prep_description_1"] = df_WIKI_DE["description"].apply(remove_artifacts)
df_WIKI_DE.loc[:, "prep_description_1"] = df_WIKI_DE["prep_description_1"].apply(unidecode.unidecode)

## 2. Remove digits, punctuation, lowercase & tokenize

df_WIKI_ES.loc[:, "QA_description"] = df_WIKI_ES["prep_description_1"].apply(lambda x: preprocess_description(x, False, False))
df_WIKI_ES.loc[:, "BERT_description"] = df_WIKI_ES["prep_description_1"].apply(lambda x: preprocess_description(x, True, False))
df_WIKI_ES.loc[:, "BOW_description"] = df_WIKI_ES["prep_description_1"].apply(lambda x: preprocess_description(x, True, True))

df_WIKI_DE.loc[:, "QA_description"] = df_WIKI_DE["prep_description_1"].apply(lambda x: preprocess_description(x, False, False))
df_WIKI_DE.loc[:, "BERT_description"] = df_WIKI_DE["prep_description_1"].apply(lambda x: preprocess_description(x, True, False))
df_WIKI_DE.loc[:, "BOW_description"] = df_WIKI_DE["prep_description_1"].apply(lambda x: preprocess_description(x, True, True))

# Save Data

df_WIKI_ES.to_excel("..//Datasets//Preprocessed Databases//WIKI_preprocessed_descriptions_ESP.xlsx", index = False)
df_WIKI_DE.to_excel("..//Datasets//Preprocessed Databases//WIKI_preprocessed_descriptions_DE.xlsx", index = False)
