==> NonEnglish Categorical Trait Prediction/02. Dataset_Creation.ipynb <==
**Notebook to create the Spanish and German Wikipedia datasets.**

This is done by combining descriptions per species and source to get one combined description for a species that might contain several WIKI categories. The descriptions are then merged with information on species traits from the Global Inventory of Floras and Traits (GIFT). 

# Libraries & Functions

'''Math & Data Libraries'''
import numpy as np
import pandas as pd

''' Miscellaneous Libraries'''
from tqdm import tqdm

# Input Description Data

## Spanish Wikipedia - WIKI_ES Dataset

df_WIKI_ES = pd.read_excel("..//Datasets//Preprocessed Databases//WIKI_preprocessed_descriptions_ESP.xlsx")

df_WIKI_ES

## German Wikipedia - WIKI_DE Dataset

df_WIKI_DE = pd.read_excel("..//Datasets//Preprocessed Databases//WIKI_preprocessed_descriptions_DE.xlsx")

df_WIKI_DE

## Statistics

print("Number of Unique Species: {}".format(df_WIKI_ES["name"].nunique()))
print("Number of Unique Description Types: {}".format(df_WIKI_ES["WIKI_id"].nunique()))

df_WIKI_ES["WIKI_id"].value_counts().describe()

print("Number of Unique Species: {}".format(df_WIKI_DE["name"].nunique()))
print("Number of Unique Description Types: {}".format(df_WIKI_DE["WIKI_id"].nunique()))

df_WIKI_DE["WIKI_id"].value_counts().describe()

# Combine Descriptions on Species

def combine_descriptions(df, df_name):
    df_new = pd.DataFrame()
    
    src_variable = "authors" if df_name=="POWO" else "Binomial Name" 

    # Adding variables for the categories and their count
    df_tmp = df.groupby(["name", src_variable])[f'{df_name}_id'].apply(';'.join).reset_index()
    df_new = df_new.assign(name = df_tmp["name"])
    df_new = df_new.assign(binname = df_tmp[src_variable])
    df_new[f'{df_name}_ids'] = df_tmp[f'{df_name}_id']
    df_new[f'{df_name}_id_N'] = [len(x.split(";")) for x in df_tmp[f'{df_name}_id']]

    # Combining the original description
    description = df["description"].apply(str)
    df = df.assign(description_str=description)
    df_tmp = df.groupby(["name", src_variable])['description_str'].apply(' ; '.join).reset_index()
    df_new = df_new.assign(description=df_tmp["description_str"])

    # Combining the QA description
    QA_description = df["QA_description"].apply(str)
    df = df.assign(QA_description_str=QA_description)
    df_tmp = df.groupby(["name", src_variable])['QA_description_str'].apply(' ; '.join).reset_index()
    df_new = df_new.assign(QA_description=df_tmp["QA_description_str"])
    
    # Combining the BERT description
    BERT_description = df["BERT_description"].apply(str)
    df = df.assign(BERT_description_str=BERT_description)
    df_tmp = df.groupby(["name", src_variable])['BERT_description_str'].apply(' '.join).reset_index()
    df_new = df_new.assign(BERT_description=df_tmp["BERT_description_str"])
    
    # Combining the BOW description
    BOW_description = df["BOW_description"].apply(str)
    df = df.assign(BOW_description_str=BOW_description)
    df_tmp = df.groupby(["name", src_variable])['BOW_description_str'].apply(' '.join).reset_index()
    df_new = df_new.assign(BOW_description=df_tmp["BOW_description_str"])

    return df_new

df_WIKI_ES_combined = combine_descriptions(df_WIKI_ES, "WIKI")
df_WIKI_DE_combined = combine_descriptions(df_WIKI_DE, "WIKI")

### Description Word Count & Character Count Analysis 

character_count_description = df_WIKI_ES_combined["BERT_description"].apply(lambda x: len(x))
df_WIKI_ES_combined = df_WIKI_ES_combined.assign(description_character_count=character_count_description)

word_count_description = df_WIKI_ES_combined["BERT_description"].apply(lambda x: len(x.split()))
df_WIKI_ES_combined = df_WIKI_ES_combined.assign(description_word_count=word_count_description)

character_count_description = df_WIKI_DE_combined["BERT_description"].apply(lambda x: len(x))
df_WIKI_DE_combined = df_WIKI_DE_combined.assign(description_character_count=character_count_description)

word_count_description = df_WIKI_DE_combined["BERT_description"].apply(lambda x: len(x.split()))
df_WIKI_DE_combined = df_WIKI_DE_combined.assign(description_word_count=word_count_description)

# Input Label Data 

## Global Inventory of Floras and Traits - GIFT Dataset

### GIFT Trait Data

df_GIFT_traits = pd.read_csv("..//Datasets//Initial Databases//GIFT_traits.csv")

df_GIFT_traits

df_GIFT_traits = df_GIFT_traits[df_GIFT_traits["restricted"]==1].drop(["restricted"], axis=1)

df_GIFT_traits

#### Filter Traits

trait_names_cat = ["Growth_form_1", "Epiphyte_1", "Climber_1", "Lifecycle_1", "Life_form_1"]
traits_cat = ["1.2.1", "1.3.1", "1.4.1", "2.1.1", "2.3.1"]

trait_names_num = ["Plant_height_max", "Leaf_length_max", "Leaf_width_max"]
traits_num = ["1.6.2", "4.6.2", "4.7.2"]

trait_dict = {trait_code: trait_name for trait_code, trait_name in zip(np.hstack((traits_cat, traits_num)), np.hstack((trait_names_cat, trait_names_num)))}

for trait_code, trait_name in zip(np.hstack((traits_cat, traits_num)), np.hstack((trait_names_cat, trait_names_num))):
    print("Trait: {}\nCode: {}\nCount {}".format(trait_name, trait_code, df_GIFT_traits["trait_ID"].value_counts()[trait_code]))
    print("------")

# Filtering Data based on selected traits - WARNING: REWRITING DATAFRAME
df_GIFT_traits = df_GIFT_traits[df_GIFT_traits["trait_ID"].apply(lambda x: x in np.hstack((traits_cat, traits_num)))]

# Drop Duplicates - WARNING: REWRITING DATAFRAME
df_GIFT_traits = df_GIFT_traits.drop_duplicates(["work_ID", "trait_ID", "trait_value", "references", "bias_by_reference", "bias_by_derivation"])

df_GIFT_traits

print("Number of Unique Species: {}".format(df_GIFT_traits["work_ID"].nunique()))
print("Average Number of Traits Per Species: {} (Max: {})".format(np.round(len(df_GIFT_traits)/df_GIFT_traits["work_ID"].nunique(), 2), len(trait_dict)))

### GIFT Species Data

df_GIFT_names = pd.read_csv("..//Datasets//Initial Databases//GIFT_names_matched.csv")

df_GIFT_names

## Combine Label Information & Convert to Wide

df_GIFT_long = df_GIFT_traits.merge(df_GIFT_names[["work_ID", "species"]], how='left', on='work_ID').drop_duplicates()

df_GIFT_long

def long_to_wide(df, trait_dict):
    df_new = pd.DataFrame(index=np.unique(df["work_ID"]))
    df_new["work_ID"] = np.unique(df["work_ID"])
    df_new["species"] = df.groupby("work_ID")["species"].aggregate(lambda x: ''.join(set(x)))
    
    for trait_code in trait_dict:
        df_tmp = df[df["trait_ID"]==trait_code]#.astype(str)
        df_tmp["agreement"] = np.round(df_tmp["agreement"], 2).astype(str)
        df_tmp["bias_by_reference"] = df_tmp["bias_by_reference"].astype(str)
        df_tmp["bias_by_derivation"] = df_tmp["bias_by_derivation"].astype(str)
        
        df_new[trait_code + "_count"] = df_tmp[df_tmp["trait_value"].notna()].groupby("work_ID")["trait_value"].count()
        df_new[trait_code] = df_tmp[df_tmp["trait_value"].notna()].groupby("work_ID")["trait_value"].aggregate(lambda x: "|".join(np.unique(np.hstack(x))))
        df_new[trait_code + "_agreement"] = df_tmp[df_tmp["trait_value"].notna()].groupby("work_ID")["agreement"].aggregate(lambda x: "|".join(np.hstack(x)))
        df_new[trait_code + "_bias_by_reference"] = df_tmp[df_tmp["trait_value"].notna()].groupby("work_ID")["bias_by_reference"].aggregate(lambda x: "|".join(np.hstack(x)))
        df_new[trait_code + "_bias_by_derivation"] = df_tmp[df_tmp["trait_value"].notna()].groupby("work_ID")["bias_by_derivation"].aggregate(lambda x: "|".join(np.hstack(x)))

    return df_new

df_GIFT_wide = long_to_wide(df_GIFT_long, trait_dict)

df_GIFT_wide

# Combine Description & Trait Data

"""Combine the WIKI and GIFT databases regards to their species name and add the wanted traits"""
def combine_description_trait_data(df, df_GIFT):
    df_GIFT["key"] = df_GIFT["species"]
    df["key"] = df["name"]
    df = pd.merge(df, df_GIFT, how = "left", on = "key")
    df.drop("key", axis=1, inplace=True)
    return df

## WIKI_GIFT

df_WIKI_GIFT_ESP = combine_description_trait_data(df_WIKI_ES_combined, df_GIFT_wide)
df_WIKI_GIFT_DE = combine_description_trait_data(df_WIKI_DE_combined, df_GIFT_wide)

## Save Data

df_WIKI_GIFT_ESP.to_excel("..//Datasets//WIKI_GIFT_ESP.xlsx", index = False)
df_WIKI_GIFT_DE.to_excel("..//Datasets//WIKI_GIFT_DE.xlsx", index = False)
